test: 당첨 로또 테스트 코드 추가
